verificacion de POST al crear y borrar

+++ views/borrar.php
@@ -1,9 +1,11 @@
 <?php
-$item = $_POST["item"];
+if ($_SERVER["REQUEST_METHOD"] == "POST") {
+    $item = $_POST["item"];
 
-require_once "model/$item.php";
+    require_once "model/$item.php";
 
-$modelo = new $item();
-$modelo->deleteOne($_POST["id"]);
+    $modelo = new $item();
+    $modelo->deleteOne($_POST["id"]);
 
-header("location:index.php?action=admin");
+    header("location:index.php?action=admin");
+}
